myerror.c AllocSprintf fails to null-pad the end of its destination string when it encounters "%s" formatted items.
The fix: strncpy an extra byte, which should be a null character at the end of each token.

Old -I  inventory lines could contain garbage characters.
example:  
3.0, 672468, 2, WWA="Watch Warning Advisory [-]B", 0-SFC, 01/23/2018 18:00, 01/23/2018 21:00, 3.00
                                               ^   unpredictable garbage character due to missing NULL
new:
3.0, 672468, 2, WWA="Watch Warning Advisory [-]", 0-SFC, 01/23/2018 18:00, 01/23/2018 21:00, 3.00

Other % formats are probably broken in this same manner. I fixed
only the %s, because that is the problem I ran into.

Frankly, Inventory should probably not even use the version of reallocSprintf found in this myerror.c
That's likely another problem.


git-svn-id: https://vlab.ncep.noaa.gov/svn/degrib/degrib@1258 98767674-7b1f-0410-897a-ff3155737c25